+++ src/const.ts
@@ -4,4 +4,11 @@ const Setting = {
 
 const CITIES = ['Paris', 'Cologne', 'Brussels', 'Amsterdam', 'Hamburg', 'Dusseldorf'] as const;
 
-export {Setting, CITIES};
+enum AppRoute {
+  MainPage = '/',
+  LoginPage = '/login',
+  FavoritesPage = '/favorites',
+  OfferPage = '/offer'
+}
+
+export {Setting, CITIES, AppRoute};
